tweaks and twerks
